Content change due to result of the EU referendum
This is because these are pages that are potentially affected but
the outcome of the EU referendum and have seen a significant spike
in visits since 23 June.

These changes were made as per this document [1] and upon request from the,
Content Team.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.