Generate regression tests artefacts (RAB)
This commit shows the current state of the regression test
artefacts as a result of the changes to the questions, responses
and expected results.

This would also affect other country routes only where Sudan
is chosen as another_country.

This has lead to the creation of the artefacts within.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

i.e With the inclusion of the aforementioned.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.