Regenerate regression tests artefacts (RAB)
This commit shows the current state of the regression test
artefacts as a result of the changes to the Content added in
previous commit.

This has led to the update of the regression tests artefacts within.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.